# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateNoticeRequest:
    properties:
      content:
        maxLength: 5000
        minLength: 5
        type: string
      expires_at:
        description: YYYY-MM-DD optional
        type: string
      pinned:
        type: boolean
      title:
        maxLength: 200
        minLength: 3
        type: string
      type:
        enum:
        - general
        - payment
        - maintenance
        - emergency
        - event
        type: string
    required:
    - content
    - title
    - type
    type: object
...
  dto.CreateTicketRequest:
    properties:
      category:
        enum:
        - plumbing
        - electrical
        - ac
        - furniture
        - other
        type: string
      description:
        maxLength: 2000
        minLength: 5
        type: string
      owner_id:
        type: string
      photos:
        items:
          type: string
        type: array
      priority:
        enum:
        - low
        - medium
        - high
        - urgent
        type: string
      room_id:
        type: string
      title:
        maxLength: 200
        minLength: 3
        type: string
    required:
    - category
    - description
    - owner_id
    - priority
    - room_id
    - title
    type: object
...
  dto.MessageListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.MessageResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.MessageResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      read:
        type: boolean
      receiver_id:
        type: string
      sender_id:
        type: string
      thread_id:
        type: string
    type: object
  dto.NoticeListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.NoticeResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.NoticeResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      owner_id:
        type: string
      pinned:
        type: boolean
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.ResolveTicketRequest:
    properties:
      notes:
        maxLength: 2000
        type: string
    type: object
...
  dto.SendMessageRequest:
    properties:
      content:
        maxLength: 2000
        minLength: 1
        type: string
      receiver_id:
        type: string
    required:
    - content
    - receiver_id
    type: object
  dto.ThreadListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ThreadResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.ThreadResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      last_at:
        type: string
      last_message:
        type: string
      participants:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  dto.TicketListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.TicketResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.TicketResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      notes:
        type: string
      owner_id:
        type: string
      photos:
        items:
          type: string
        type: array
      priority:
        type: string
      resolved_at:
        type: string
      room_id:
        type: string
      status:
        type: string
      tenant_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  dto.UpdateNoticeRequest:
    properties:
      content:
        maxLength: 5000
        minLength: 5
        type: string
      expires_at:
        description: YYYY-MM-DD or null to clear
        type: string
      pinned:
        type: boolean
      title:
        maxLength: 200
        minLength: 3
        type: string
      type:
        enum:
        - general
        - payment
        - maintenance
        - emergency
        - event
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
  description: Production-ready Go REST API starter with hexagonal architecture.
  title: Go Starter API
  version: "1.0"
paths:
...
  /api/v1/maintenance:
    get:
      parameters:
      - description: Filter by room
        in: query
        name: room_id
        type: string
      - description: Filter by tenant
        in: query
        name: tenant_id
        type: string
      - description: Filter by owner
        in: query
        name: owner_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by priority
        in: query
        name: priority
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TicketListResponse'
      security:
      - BearerAuth: []
      summary: List maintenance tickets
      tags:
      - maintenance
    post:
      consumes:
      - application/json
      parameters:
      - description: Ticket data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TicketResponse'
      security:
      - BearerAuth: []
      summary: Open a maintenance ticket
      tags:
      - maintenance
  /api/v1/maintenance/{id}:
    get:
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TicketResponse'
      security:
      - BearerAuth: []
      summary: Get maintenance ticket
      tags:
      - maintenance
  /api/v1/maintenance/{id}/close:
    post:
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TicketResponse'
      security:
      - BearerAuth: []
      summary: Close a resolved ticket (tenant confirmation)
      tags:
      - maintenance
  /api/v1/maintenance/{id}/resolve:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: Resolution notes
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.ResolveTicketRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TicketResponse'
      security:
      - BearerAuth: []
      summary: Resolve a maintenance ticket
      tags:
      - maintenance
  /api/v1/maintenance/{id}/start:
    post:
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TicketResponse'
      security:
      - BearerAuth: []
      summary: Start working on a ticket
      tags:
      - maintenance
  /api/v1/messages:
    post:
      consumes:
      - application/json
      parameters:
      - description: Message data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Send a message
      tags:
      - messages
  /api/v1/messages/threads:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ThreadListResponse'
      security:
      - BearerAuth: []
      summary: List conversation threads
      tags:
      - messages
  /api/v1/messages/threads/{threadId}:
    get:
      parameters:
      - description: Thread ID
        in: path
        name: threadId
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageListResponse'
      security:
      - BearerAuth: []
      summary: Get messages in a thread
      tags:
      - messages
  /api/v1/messages/threads/{threadId}/read:
    post:
      parameters:
      - description: Thread ID
        in: path
        name: threadId
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Mark thread messages as read
      tags:
      - messages
  /api/v1/notices:
    get:
      parameters:
      - description: Filter by owner
        in: query
        name: owner_id
        type: string
      - description: Filter by type
        in: query
        name: type
        type: string
      - description: Show pinned only
        in: query
        name: pinned_only
        type: boolean
      - description: Exclude expired
        in: query
        name: active_only
        type: boolean
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NoticeListResponse'
      summary: List notices
      tags:
      - notices
    post:
      consumes:
      - application/json
      parameters:
      - description: Notice data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateNoticeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.NoticeResponse'
      security:
      - BearerAuth: []
      summary: Post an announcement
      tags:
      - notices
  /api/v1/notices/{id}:
    delete:
      parameters:
      - description: Notice ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a notice
      tags:
      - notices
    get:
      parameters:
      - description: Notice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NoticeResponse'
      summary: Get a notice
      tags:
      - notices
    put:
      consumes:
      - application/json
      parameters:
      - description: Notice ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateNoticeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NoticeResponse'
      security:
      - BearerAuth: []
      summary: Update a notice
      tags:
      - notices
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"